+++ src/main/java/cc/cassian/raspberry/registry/RaspberryTags.java
@@ -20,6 +20,7 @@ public class RaspberryTags {
     public static final TagKey<EntityType<?>> WORM_SEEKERS = createEntityTypeTag( "worm_seekers");
 
     public static final TagKey<Block> SHEARS_SHOULD_MINE = createBlockTag("shears_should_mine");
+    public static final TagKey<Block> SHEARS_SHOULD_USE = createBlockTag("shears_should_use");
 
     public static TagKey<Block> createBlockTag(String id) {
         return TagKey.create(ForgeRegistries.BLOCKS.getRegistryKey(), new ResourceLocation(MOD_ID, id));